3.17-v2

+++ finetune.py
@@ -545,18 +545,18 @@ def finetune(args, tokenizer: AutoTokenizer, model: deepspeed.DeepSpeedEngine, o
             #     dist.barrier()
 
             # Evaluation
-            if args.eval_interval: 
+            if args.eval_interval and global_step % args.eval_interval == 0 and step % args.gradient_accumulation_steps == 0:
                 eval_result = evaluate(args, tokenizer, model, dataset["dev"], "dev", epoch, device)
 
                 for ckpt_key, ckpt_path in last_ckpt_path.items():
                     if ckpt_path is not None and os.path.exists(ckpt_path):
                         shutil.rmtree(ckpt_path)
-                if eval_result['avg_loss'] > np.max(eval_results['avg_loss']):
+                if eval_result['avg_loss'] >= np.max(eval_results['avg_loss']):
                     ckpt_name = f'avg_loss_{str(epoch)}'
                     save_ckpt(args, ckpt_name, model, tokenizer)
                     last_ckpt_path['avg_loss'] = os.path.join(args.save, ckpt_name)
                     
-                if eval_result['rougeL'] > np.max(eval_results['rougeL']):
+                if eval_result['rougeL'] >= np.max(eval_results['rougeL']):
                     ckpt_name = f'rougeL_{str(epoch)}'
                     save_ckpt(args, ckpt_name, model, tokenizer)
                     last_ckpt_path['rougeL'] = os.path.join(args.save, ckpt_name)
